pyproject.toml
[tool.poetry]
name = "skrotowiec"
version = "0.1.0"
description = "Zadanie rekrutacyjne - Backend Developer"
authors = ["Dariusz Rybi <dariusz.rybi@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.18.1"
pylint = "^3.3.8"

[dependency-groups]
test = [
		"pytest",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "django.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true

[tool.pylint."messages control"]
disable = ["raw-checker-failed", "bad-inline-option", "locally-disabled", "file-ignored", "suppressed-message", "useless-suppression", "deprecated-pragma", "use-symbolic-message-instead", "use-implicit-booleaness-not-comparison-to-string", "use-implicit-booleaness-not-comparison-to-zero", "missing-module-docstring", "missing-class-docstring", "too-few-public-methods", "misplaced-bare-raise"]
